Introduce support for loading cap info from term.h
This commit adds support for loading a subset of the
capabilities information from term.h instead a caps file.
This is useful when you can't get to a capabilities table
file.  However, only the variable name and number are loaded,
and you must pass `use_variable_names=True` to `TermInfo`.